refactor: :recycle: replace custom day of year calculations with TimeManager methods and streamline related logic
