+++ src/settings/settings.ts
@@ -81,6 +81,7 @@ export class SettingsManager {
   isBlockPersistence = false;
   /** Center on a satellite when it is selected. */
   isFocusOnSatelliteWhenSelected = true;
+  isUseJdayOnTopMenu = false;
 
 
   static preserveSettings() {
